+++ M2/Macaulay2/e/imonorder.cpp
@@ -10,7 +10,7 @@
 #  include <malloc.h>
 #endif
 
-#include "ExponentVector.hpp"
+#include "monomials/ExponentVector.hpp"
 #include "monomials/overflow.hpp"
 
 std::vector<bool> laurentVariables(const MonomialOrder* mo)
