+++ M2/Macaulay2/e/schreyer-resolutions/res-monomial-types.hpp
@@ -5,8 +5,8 @@
 
 #include <cstdint>
 
-#include "ExponentList.hpp"
-#include "ExponentVector.hpp"
+#include "monomials/ExponentList.hpp"
+#include "monomials/ExponentVector.hpp"
 
 enum class MonomialOrderingType { Lex, GRevLex, Weights };
